Refactored GraphQl
