+++ api/graphql/schema.py
@@ -1,0 +1,6 @@
+import graphene
+
+from .queries import Query
+from .mutations import Mutation
+
+schema = graphene.Schema(query=Query, mutation=Mutation)
